+++ i18n/messages/resources/hi.json
@@ -1,5 +1,7 @@
 {
   "Resources": {
+    "backToConversations": "Conversations par vaapas jaaen",
+    "backToShorts": "Clips par vaapas jaaen",
     "conversations": "Bātcīt",
     "shorts": "Clips",
     "completed": "Pūra",
